requirements.txt
# === Core Inference ===
onnxruntime>=1.10.0,<2
ai-edge-litert>=2.0.2,<3
numpy
scipy>=1.3,<2
requests>=2.0,<3
tqdm>=4.0,<5.0
scikit-learn>=1,<2
pyyaml>=6.0,<7

# === Training ===
torch>=1.13.1,<3
torchaudio>=0.13.1,<1
torchinfo>=1.8.0,<2
torchmetrics>=0.11.4,<1
speechbrain>=0.5.14,<1
audiomentations>=0.30.0,<1
torch-audiomentations>=0.11.0,<1
acoustics>=0.2.6,<1
mutagen>=1.46.0,<2
pronouncing>=0.2.0,<1
deep-phonemizer==0.0.19
datasets>=2.14.4,<3

# === Model Conversion ===
onnx==1.14.0
onnx2tf>=1.17.0
onnx_tf==1.10.0
tensorflow-cpu==2.8.1
tensorflow_probability==0.16.0
protobuf>=3.20,<4

# === TTS ===
dashscope
edge-tts

# === Environment ===
python-dotenv
